+++ packages/components/MenuButton/src/MenuButton.tsx
@@ -6,7 +6,7 @@ import { IUseComposeStyling, compose } from '@uifabricshared/foundation-compose'
 import { mergeSettings } from '@uifabricshared/foundation-settings';
 import { ISlots, withSlots } from '@uifabricshared/foundation-composable';
 import { backgroundColorTokens, borderTokens } from '@fluentui-react-native/tokens';
-import { Svg, Path } from 'react-native-svg';
+import { SvgXml } from 'react-native-svg';
 import { defaultIconColor, primaryIconColor } from './MenuButton.style';
 
 import {
@@ -34,7 +34,7 @@ export const MenuButton = compose<MenuButtonType>({
       setShowContextualMenu(!showContextualMenu);
     }, [showContextualMenu, setShowContextualMenu]);
 
-    const menuItemsUpdated = menuItems.map((item) => {
+    const menuItemsUpdated = menuItems.map(item => {
       if (item.hasSubmenu) {
         const [showSubmenu, setShowSubmenu] = useState(false);
 
@@ -88,12 +88,6 @@ export const MenuButton = compose<MenuButtonType>({
       contextualMenuItems: {
         menuItems: menuItemsUpdated,
       },
-      chevronSvg: {
-        width: '12',
-        height: '16',
-        viewBox: '0 0 11 6',
-        color: primary ? primaryIconColor : defaultIconColor,
-      },
     });
 
     return { slotProps, state };
@@ -104,7 +98,7 @@ export const MenuButton = compose<MenuButtonType>({
     primaryButton: { slotType: PrimaryButton as React.ComponentType },
     contextualMenu: { slotType: ContextualMenu as React.ComponentType },
     contextualMenuItems: React.Fragment,
-    chevronSvg: Svg,
+    chevronSvg: SvgXml,
   },
   styles: {
     contextualMenu: [backgroundColorTokens, borderTokens],
@@ -117,20 +111,26 @@ export const MenuButton = compose<MenuButtonType>({
     const context = renderData.state!.context;
     const menuItems = renderData.slotProps!.contextualMenuItems ? renderData.slotProps!.contextualMenuItems.menuItems : [];
 
-    const chevronIcon = () => {
-      const chevronPath = `M0.646447 0.646447C0.841709 0.451184 1.15829 0.451184 1.35355 0.646447L5.5 4.79289L9.64645 0.646447C9.84171 0.451185 10.1583 0.451185 10.3536 0.646447C10.5488 0.841709 10.5488 1.15829 10.3536 1.35355L5.85355 5.85355C5.65829 6.04882 5.34171 6.04882 5.14645 5.85355L0.646447 1.35355C0.451184 1.15829 0.451184 0.841709 0.646447 0.646447Z`;
-      return (
-        <Slots.chevronSvg>
-          <Path fill="currentColor" d={chevronPath} />
-        </Slots.chevronSvg>
-      );
-    };
+    const chevronColor = context.primary ? primaryIconColor : defaultIconColor;
+    const chevronXml = `
+    <svg width="12" height="16" viewBox="0 0 11 6" color=${chevronColor}>
+      <path fill='currentColor' d='M0.646447 0.646447C0.841709 0.451184 1.15829 0.451184 1.35355 0.646447L5.5 4.79289L9.64645 0.646447C9.84171 0.451185 10.1583 0.451185 10.3536 0.646447C10.5488 0.841709 10.5488 1.15829 10.3536 1.35355L5.85355 5.85355C5.65829 6.04882 5.34171 6.04882 5.14645 5.85355L0.646447 1.35355C0.451184 1.15829 0.451184 0.841709 0.646447 0.646447Z' />
+    </svg>`;
+
     return (
       <Slots.root>
-        {context.primary ? <Slots.primaryButton>{chevronIcon()}</Slots.primaryButton> : <Slots.button>{chevronIcon()}</Slots.button>}
+        {context.primary ? (
+          <Slots.primaryButton>
+            <Slots.chevronSvg xml={chevronXml} />
+          </Slots.primaryButton>
+        ) : (
+          <Slots.button>
+            <Slots.chevronSvg xml={chevronXml} />
+          </Slots.button>
+        )}
         {context.showContextualMenu && (
           <Slots.contextualMenu>
-            {menuItems.map((menuItem) => {
+            {menuItems.map(menuItem => {
               const { hasSubmenu, submenuProps, showSubmenu, componentRef, submenuItems, ...items } = menuItem;
 
               return hasSubmenu && submenuItems ? (
@@ -138,7 +138,7 @@ export const MenuButton = compose<MenuButtonType>({
                   <SubmenuItem componentRef={componentRef} {...items} />
                   {showSubmenu && (
                     <Submenu target={componentRef} {...submenuProps}>
-                      {submenuItems.map((submenuItem) => (
+                      {submenuItems.map(submenuItem => (
                         <ContextualMenuItem key={submenuItem.itemKey} {...submenuItem} />
                       ))}
                     </Submenu>